refactor: inline single-caller build* helpers (#7745)
Three exported build* functions had exactly one non-test production
caller, banned by CLAUDE.md's "Discouraged Factory Patterns" section:

- buildFileContextFromTaskState (agent/runtime/textEnhancement.ts) ->
  inlined as a private method on its single caller,
  ProgressFollowUpPolishController.
- buildToolDashboardTerminalAction (controllers/settingsView/ToolDashboardData.ts)
  -> moved into SettingsViewMessageHandler.ts (its only caller; the
  desktop host never calls it, so it wasn't actually cross-host shared
  despite living in the shared settingsView controllers directory).
- buildClaudeAgentWorkspaceOptions (tools/claudeAgentConfig.ts) ->
  inlined at its single call site in claudeAgent.ts, which now calls
  the still-shared buildAgentWorkspaceOptions directly and maps its
  workingDirectory/additionalDirectories fields to the claude-agent-sdk's
  cwd/additionalDirectories naming inline.
